fix(issu16): Fixed bug on test of mongoengine signals.
Closes #16
